+++ apps/webapp/src/pages/main.tsx
@@ -4,7 +4,7 @@ import { ConfigProvider } from '../modules/config/context/ConfigContext';
 import { App } from './App';
 import { ErrorBoundary } from '../modules/layout/components/ErrorBoundary';
 
-import '@jetstreamgg/widgets/style.css';
+import '@jetstreamgg/widgets/globals.css';
 import '../globals.css';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(