+++ apps/webapp/vite.config.ts
@@ -5,10 +5,14 @@ import { configDefaults } from 'vitest/config';
 import { lingui } from '@lingui/vite-plugin';
 import tailwindcss from 'tailwindcss';
 import simpleHtmlPlugin from 'vite-plugin-simple-html';
-import { nodePolyfills } from 'vite-plugin-node-polyfills';
+
+enum modeEnum {
+  development = 'development',
+  production = 'production'
+}
 
 // https://vitejs.dev/config/
-export default ({ mode }: { mode: string }) => {
+export default ({ mode }: { mode: modeEnum }) => {
   process.env = { ...process.env, ...loadEnv(mode, process.cwd()) };
 
   const RPC_PROVIDER_MAINNET = process.env.VITE_RPC_PROVIDER_MAINNET || '';
@@ -91,7 +95,16 @@ export default ({ mode }: { mode: string }) => {
     },
     resolve: {
       alias: {
-        '@': path.resolve(__dirname, './src')
+        '@': path.resolve(__dirname, './src'),
+        // If we're in dev mode, alias the packages to their local TypeScript source code for faster HMR
+        ...(mode === modeEnum.development
+          ? {
+              '@jetstreamgg/hooks': path.resolve(__dirname, '../../packages/hooks/src'),
+              '@jetstreamgg/utils': path.resolve(__dirname, '../../packages/utils/src'),
+              '@jetstreamgg/widgets': path.resolve(__dirname, '../../packages/widgets/src'),
+              '@widgets': path.resolve(__dirname, '../../packages/widgets/src')
+            }
+          : {})
       }
     },
     optimizeDeps: {
@@ -113,12 +126,6 @@ export default ({ mode }: { mode: string }) => {
             }
           ]
         }
-      }),
-      nodePolyfills({
-        globals: {
-          process: false
-        },
-        include: ['buffer']
       }),
       react({
         plugins: [['@lingui/swc-plugin', {}]]