+++ packages/widgets/src/components/ui/tabs.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as TabsPrimitive from '@radix-ui/react-tabs';
 import { cva, type VariantProps } from 'class-variance-authority';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@widgets/lib/utils';
 
 const Tabs = TabsPrimitive.Root;